backend/requirements.txt
# Certificate Generation System - Dependencies

# Core Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Data Validation
pydantic[email]==2.5.3
pydantic-settings==2.1.0
jsonschema==4.21.1

# Authentication
pyjwt==2.8.0

# Database
sqlalchemy==2.0.25
asyncpg==0.29.0
greenlet==3.0.3

# PDF Rendering
weasyprint==59.0
pydyf==0.8.0
jinja2==3.1.3

# Image Conversion
pdf2image==1.17.0
pillow==10.2.0

# Storage
boto3==1.34.25

# Utilities
python-dotenv==1.0.0

# Development
httpx==0.26.0
pytest==7.4.4
pytest-asyncio==0.23.3
